+++ src/main/java/com/l2jserver/gameserver/network/serverpackets/GMViewCharacterInfo.java
@@ -90,18 +90,18 @@ public class GMViewCharacterInfo extends L2GameServerPacket
 		
 		writeD(_activeChar.getInventory().getTalismanSlots()); // CT2.3
 		writeD(_activeChar.getInventory().canEquipCloak() ? 1 : 0); // CT2.3
-		writeD(_activeChar.getPAtk(null));
-		writeD(_activeChar.getPAtkSpd());
-		writeD(_activeChar.getPDef(null));
+		writeD((int) _activeChar.getPAtk(null));
+		writeD((int) _activeChar.getPAtkSpd());
+		writeD((int) _activeChar.getPDef(null));
 		writeD(_activeChar.getEvasionRate(null));
 		writeD(_activeChar.getAccuracy());
 		writeD(_activeChar.getCriticalHit(null, null));
-		writeD(_activeChar.getMAtk(null, null));
+		writeD((int) _activeChar.getMAtk(null, null));
 		
 		writeD(_activeChar.getMAtkSpd());
-		writeD(_activeChar.getPAtkSpd());
+		writeD((int) _activeChar.getPAtkSpd());
 		
-		writeD(_activeChar.getMDef(null, null));
+		writeD((int) _activeChar.getMDef(null, null));
 		
 		writeD(_activeChar.getPvpFlag()); // 0-non-pvp 1-pvp = violett name
 		writeD(_activeChar.getKarma());
